% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dataset_doc.R
\name{subspace_dataset}
\alias{subspace_dataset}
\title{Synthetic Subspace Clustering Dataset}
\description{
This is a synthetic dataset to demonstrate the subspace clustering algorithms
in the package \emph{subspace}
}

